backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
rembg==2.0.56
onnxruntime-gpu; sys_platform == 'linux'
onnxruntime; sys_platform == 'darwin'
torch
torchvision
imageio
imageio-ffmpeg
trimesh
scipy
tqdm
